Move OSC code to CardinalOSC.{cpp,hpp}
