+++ src/CardinalRemote/Makefile
@@ -329,6 +329,7 @@ BUILD_CXX_FLAGS += -DCARDINAL_PLUGIN_PREFIX='"$(PREFIX)"'
 FILES  = main.cpp
 FILES += RemoteUI.cpp
 FILES += CardinalCommon.cpp
+FILES += CardinalOSC.cpp
 FILES += CardinalRemote.cpp
 FILES += common.cpp
 FILES += glfw.cpp
